pyproject.toml
[project]
name = "deeperlang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.302",
    "langroid[metaphor]>=0.49.0",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
    "typer>=0.15.2",
]
